Fix api_proxy delete method to only succeed when proxy is deleted
Previously, when the delete method fails to delete the proxy it still claims to have succeeded and removes the proxy from state.
For example, when a proxy has a deployment associated with it, it cannot be deleted.
This was due to a small bug in the delete method.

Added a test to check this behaviour.